Price Implementation/vapi_service_source/vapi_service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
stripe==6.6.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
psutil==5.9.5
email-validator==2.0.0
bcrypt==4.0.1
redis==4.6.0
celery==5.3.1
APScheduler==3.10.4
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
dnspython==2.4.2
cryptography==41.0.4
PyJWT==2.8.0
marshmallow==3.20.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
urllib3==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4

